backend/requirements/test.txt
-r development.txt
coverage==5.5
pytest==6.2.4
pytest-cov==2.12.0
pytest-django==4.3.0
